Robot tests: wait until overlay is closed.
Use the correct selector to check for the overlay: it is no longer `div.overlay`, but `div.modal-wrapper`.
Wait only 8 seconds instead of 60 for this check to be true.
Also use 'Wait For Then Click Element' for good measure.

This is an attempt to make robot tests more stable
I still sometimes see a test fail, becase the popup/modal/overlay is still open after clicking a button.  This might fix it.

Update: no it does not.  But the updated checks at least make it clear that a popup is still open, which is the reason why some other expected text is not visible.