Fixes random failure in robot test `Scenario: Add a fieldSet and move a field into this fieldset`
Fixes the error:

Element states '['detached']' (list) should contain '['visible']' (list)

This error occurred on the line:
`Wait For Condition Element States //div[contains(@class,"modal-dialog")] contains visible`

of `Click Modal Link` keyword.

What happened was that the test tried to click on the element before
the click event was attached to the element. Now we ensure that the
event was attached before clicking.